+++ components/merkledb/src/indexes/mod.rs
@@ -1,0 +1,32 @@
+// Copyright 2019 The Exonum Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+pub use self::{
+    entry::Entry, group::Group, key_set_index::KeySetIndex, list_index::ListIndex,
+    map_index::MapIndex, proof_entry::ProofEntry, sparse_list_index::SparseListIndex,
+    value_set_index::ValueSetIndex,
+};
+
+mod entry;
+mod group;
+mod proof_entry;
+
+mod key_set_index;
+mod list_index;
+mod map_index;
+mod sparse_list_index;
+mod value_set_index;
+
+pub mod proof_list_index;
+pub mod proof_map_index;